expose getRawBody from kit/http
