+++ .changeset/tiny-candles-repeat.md
@@ -1,0 +1,5 @@
+---
+'@sveltejs/kit': patch
+---
+
+Expose getRawBody from kit/http
